Add files via upload
New version of Cimon, having trouble though when I want to go to the next turn how to go about doing that because the current setup is that the game is pretty much run through the updatePlayersSequenceThisTurn() function which is only called when a button is pressed. I'm not sure how to make it so that the cimonGame() function inside view controller class to control the game and allow for the input to be used or prohibited.