Add files via upload
I combined the functions for each color button pressed into one function that sends the title of button which is the color. Will have to adjust later because in the end the buttons won't have any text on them but for now it's fine, just trying to streamline the code.