fixed overlapping button animations
i was walking down main st last night and it dawned on me that to prevent the button animations from overlapping one another the function that performs the button animations would have to be called recursively in the completion: parameter of UIButton.animate. all that's left to do now is load a new view when the player loses the game which asks them if they'd like to play again.